好奇怪，cnt1没有增加，难道在定时器中不能改数据?

+++ SYSTEM/delay/delay.c
@@ -53,7 +53,7 @@ void delay_systick_handler(void)
     }
 }
 
-u32 delay_get_ms(void)
+uint64_t delay_get_ms(void)
 {
     return g_tick_ms;
 }
